predict/code/requirements.txt
numpy==1.14.3
pandas==0.24.2
Pillow==5.2.0
requests==2.20.0